+++ packages/merge-queue/src/types.ts
@@ -1,0 +1,41 @@
+// Copyright 2022 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+export interface Queue {
+  repoFullName: string; // e.g. "googleapis/repo-automation-bots"
+  pullRequests: number[];
+}
+
+export interface CallbackCorePayload {
+  task_type: string;
+  pr_number: number;
+  merge_effort_started_at?: string;
+}
+
+export interface CallbackPayload extends CallbackCorePayload {
+  repository: {
+    name: string;
+    full_name: string;
+    owner: {
+      login: string;
+      name: string;
+    };
+  };
+  organization: {
+    login: string;
+  };
+  installation: {
+    id: number;
+  };
+}
